코드값 백엔드에서 가져오기

+++ backend/src/main/java/com/tournet/tournetERP/common/controller/ComCodeController.java
@@ -110,7 +110,7 @@ public class ComCodeController {
         modifyingUser.setEmpUuid(userDetails.getEmpUuid());
         /**E: 수정자 정보**/
 
-        int id = comcode.getCodeUuid();
+        long id = (long) comcode.getCodeUuid();
 
         Optional<ComCode> currentComCode = comCodeRepository.findByCodeUuid(id);
